+++ domain/src/test/kotlin/fr/vbrosseau/tailscaleautorules/domain/usecase/DescribeTunnelStatusUseCaseTest.kt
@@ -57,6 +57,7 @@ class DescribeTunnelStatusUseCaseTest {
                     networkObserver = observer,
                     evaluateRules = evaluateRules,
                     clock = clock,
+                    sessionAttestation = SessionAttestation(clock),
                 ),
             journalRepository = journal,
             controller = controller,